Move getLogin from github to vcs package object
