+++ modules/core/src/main/scala/org/scalasteward/core/vcs/package.scala
@@ -17,10 +17,10 @@
 package org.scalasteward.core
 
 import org.scalasteward.core.application.Config
-import org.scalasteward.core.vcs.data.Repo
 import org.scalasteward.core.model.Update
+import org.scalasteward.core.vcs.data.Repo
 
-package object github {
+package object vcs {
 
   def getLogin(config: Config, repo: Repo): String =
     if (config.doNotFork) repo.owner else config.gitHubLogin
