+++ modules/core/src/test/scala/org/scalasteward/core/vcs/VCSPackageTest.scala
@@ -1,10 +1,10 @@
-package org.scalasteward.core.github
+package org.scalasteward.core.vcs
 
-import org.scalasteward.core.vcs.data.Repo
 import org.scalasteward.core.mock.MockContext.config
+import org.scalasteward.core.vcs.data.Repo
 import org.scalatest.{FunSuite, Matchers}
 
-class GithubPackageTest extends FunSuite with Matchers {
+class VCSPackageTest extends FunSuite with Matchers {
   val repo = Repo("fthomas", "datapackage")
 
   test("github login for fork enabled configuration") {
